Update remove users migration

+++ db/migrate/20181108234254_remove_users_from_main_db.rb
@@ -1,5 +1,6 @@
 class RemoveUsersFromMainDb < ActiveRecord::Migration[5.2]
   def change
+    return if !table_exists? :admin_users
     drop_table :admin_users
   end
 end
